requirements.txt
streamlit
streamlit_extras
requests
regex
geopy
geopandas
leafmap
owslib
matplotlib
SQLAlchemy 
mysqlclient
openpyxl
setuptools
streamlit-shadcn-ui
pygwalker
st-circular-progress
seaborn
